Add in new cli modules
