+++ nlpsandboxclient/cli/__main__.py
@@ -2,7 +2,7 @@
 import click
 
 from .. import __version__
-from . import community, evaluate
+from . import annotator_cli, community, datanode_cli, evaluate
 
 
 @click.group()
@@ -13,5 +13,7 @@ def cli():
 
 def main():
     cli.add_command(community.cli)
-    cli.add_command(evaluate.cli)
+    cli.add_command(evaluate.evaluate_prediction)
+    cli.add_command(datanode_cli.cli)
+    cli.add_command(annotator_cli.cli)
     cli()
